Consolidate the two Claimer contracts into one that runs on all three legs
PR #5412 landed a Claimer contract concurrently with this branch. Neither
dominated: theirs had four cases plus an aggregator, mine eight; theirs pinned
the refusal PROSE, mine the claim pools, the configured active statuses and the
bare-row promise.

The difference that decided it is the wiring. Theirs ran through
conformance.RunAll, whose Factory is `func(*testing.T) storage.DoltStorage` —
and a unit-of-work provider is not one, so only the two store legs could reach
it. Those two share internal/storage/issueops.ExecuteClaim. So every promise it
pinned was compared between two implementations that already run the same code,
and never against the one backend that does not: uow, whose Claim lives in
internal/storage/domain/db.

This keeps the three-leg contract, absorbs their refusal-message case into it,
deletes the audit-suite copy, and removes the RunAll hook with a comment saying
why the role cannot be tested from there.

Their case is the best thing in either file and is now checked on uow for the
first time: the open-but-assigned refusal OMITS the "claimed by" fragment on
purpose while the typed *ClaimConflictError still carries the holder, so a
caller reading prose learns less than one reading the struct. That asymmetry is
deliberate, and nothing outside that case says so.

Nine cases, three legs: 19 subtests at each store leg and 34 at uow, green.
